+++ src/main_broken.cpp
@@ -1,9 +1,9 @@
 #include <iostream>
 
-double deviation( double* a, int n);
+double deviation( int* a, int n);
 
 int main()
 {
-	  double a[4] = {1,2,3,4};
+	  int a[4] = {1,2,3,4};
 		std::cout << deviation(a, 4) << std::endl;
 }
